Move legacy_proto to protocol
