+++ client/network/src/protocol/legacy_proto/tests.rs
@@ -26,7 +26,7 @@ use libp2p::{PeerId, Multiaddr, Transport};
 use rand::seq::SliceRandom;
 use std::{io, time::Duration, time::Instant};
 use crate::message::Message;
-use crate::legacy_proto::{LegacyProto, LegacyProtoOut};
+use crate::protocol::legacy_proto::{LegacyProto, LegacyProtoOut};
 use test_client::runtime::Block;
 
 /// Builds two nodes that have each other as bootstrap nodes.
